+++ packages/path_provider/path_provider/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 2.0.8
+
+* Updates example app Android compileSdkVersion to 31.
+* Removes obsolete manual registration of Windows and Linux implementations.
+
+## 2.0.7
+
+* Moved Android and iOS implementations to federated packages.
+
+## 2.0.6
+
+* Added support for Background Platform Channels on Android when it is
+  available.
+
 ## 2.0.5
 
 * Update minimum Flutter SDK to 2.5 and iOS deployment target to 9.0.