+++ packages/quick_actions/quick_actions_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
-# 1.0.0
+## NEXT
+
+* Updates code for analyzer changes.
+
+## 1.0.0
 
 * Initial release of quick_actions_platform_interface