+++ packages/video_player/video_player/CHANGELOG.md
@@ -1,3 +1,30 @@
+## 2.2.10
+
+* iOS: Updates texture on `seekTo`.
+
+## 2.2.9
+
+* Adds compatibility with `video_player_platform_interface` 5.0, which does not
+  include non-dev test dependencies.
+
+## 2.2.8
+
+* Changes the way the `VideoPlayerPlatform` instance is cached in the
+  controller, so that it's no longer impossible to change after the first use.
+* Updates unit tests to be self-contained.
+* Fixes integration tests.
+* Updates Android compileSdkVersion to 31.
+* Fixes a flaky integration test.
+
+## 2.2.7
+
+* Fixes a regression where dragging a [VideoProgressIndicator] while playing
+  would restart playback from the start of the video.
+
+## 2.2.6
+
+* Initialize player when size and duration become available on iOS
+
 ## 2.2.5
 
 * Support to closed caption WebVTT format added.